Updated NetworkManager

+++ Wallet App/NewTransaction/CustomButton.swift
@@ -11,14 +11,14 @@ class CustomButton: UIButton {
     
     var delegate: CurrencyButtonDelegate?
     
-    func configure(_ title: String, _ background: UIColor, _ cornerRadius: Int) {
+    func configure(_ title: String, _ background: UIColor, _ cornerRadius: CGFloat) {
         setTitle(title, for: .normal)
         setTitleColor(UIColor(named: "text"), for: .normal)
         backgroundColor = background
         layer.cornerRadius = cornerRadius
         translatesAutoresizingMaskIntoConstraints = false
     }
-    }
+}
 
 protocol CurrencyButtonDelegate {
